L13/requirements.txt
# Multi-Agent Translation Turing Machine
# Dependencies

# Embeddings and NLP
sentence-transformers>=2.2.0

# Vector operations
numpy>=1.24.0
scikit-learn>=1.3.0

# API clients
anthropic>=0.18.0

# Visualization
matplotlib>=3.7.0

# Environment variables
python-dotenv>=1.0.0
